Refactored uniqueSelectors into a common enum file, and content into commonStaticText.ts
